+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Removed `gantry cluster` commands.
 
+### Fixed
+
+- Don't prompt for budget when workspace has a default budget associate with it.
+
 ## [v1.17.0](https://github.com/allenai/beaker-gantry/releases/tag/v1.17.0) - 2025-04-28
 
 ### Added