+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v2.0.0](https://github.com/allenai/beaker-gantry/releases/tag/v2.0.0) - 2025-05-01
+
 ### Changed
 
 - Python >= 3.10 required.